Implement data conversion and filtering classes for COVID-19 analysis
Created the Converter class, which handles the conversion of raw data to data that can be easily manipulated by the user - thus, converting them from CSV to Array. In the parent Data class, the data are converted from an array format to specific data of day and number of deaths based on a specific country. In the child Country class, a country is specified so that the parent Data class returns specific data for that country.